+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 - Add new `-q, --quiet` flag, see #303 (@Asha20)
 
+- Add new `--no-ignore-parent` flag, see #787 (@will459)
+
 ## Bugfixes
 
 - Set default path separator to `/` in MSYS, see #537 and #730 (@aswild)